+++ src/hooks/useNYPLBreakpoints.ts
@@ -14,17 +14,20 @@ const useNYPLBreakpoints = () => {
     isLargerThanMedium: false,
     isLargerThanLarge: false,
     isLargerThanXLarge: false,
+    isSmallerThanMedium: false,
   });
   const [
     isLargerThanSmall,
     isLargerThanMedium,
     isLargerThanLarge,
     isLargerThanXLarge,
+    isSmallerThanMedium,
   ] = useMediaQuery([
     "(min-width: 320px)",
     "(min-width: 600px)",
     "(min-width: 960px)",
     "(min-width: 1280px)",
+    "(max-width: 599px)",
   ]);
 
   useEffect(() => {
@@ -33,12 +36,14 @@ const useNYPLBreakpoints = () => {
       isLargerThanMedium,
       isLargerThanLarge,
       isLargerThanXLarge,
+      isSmallerThanMedium,
     });
   }, [
     isLargerThanSmall,
     isLargerThanMedium,
     isLargerThanLarge,
     isLargerThanXLarge,
+    isSmallerThanMedium,
   ]);
 
   return {
@@ -49,6 +54,11 @@ const useNYPLBreakpoints = () => {
     isLargerThanMobile: layoutSize.isLargerThanMedium,
     isLargerThanLarge: layoutSize.isLargerThanLarge,
     isLargerThanXLarge: layoutSize.isLargerThanXLarge,
+    // The values use a desktop-first approach. In an effort to mitigate the
+    // flashing font size issue in the Heading responsive font sizes, the DS
+    // will use a desktop-first style approach.
+    isSmallerThanMedium: layoutSize.isSmallerThanMedium,
+    isSmallerThanDesktop: layoutSize.isSmallerThanMedium,
   };
 };