+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Updates
+
+- Updates the `Heading` component to use a desktop-first approach to implement the responsive `font-size` styles.
+
 ## 2.1.4 (January 4, 2024)
 
 ### Adds